refactor(chat): extract canonical turn admission boundaries
